:recycle: Refactor BoosterPackCard in its own file
